// src/api/types.gen.ts
export type RegisterRequest = {
    email: string;
    password: string;
    first_name: string;
    last_name: string;
    role: 'ADMIN' | 'TEACHER';
};

export type LoginRequest = {
    email: string;
    password: string;
};

export type AuthResponse = {
    access_token: string;
    token_type: string;
    user: UserResponse;
};

export type UserResponse = {
    id: number;
    email: string;
    first_name: string;
    last_name: string;
    role: 'ADMIN' | 'TEACHER';
    status: 'ACTIVE' | 'BLOCKED';
    created_at: string;
};